src/third_party/blink/web_tests/external/wpt/webrtc/RTCRtpParameters-degradationPreference-expected.txt

This is a testharness.js-based test.
FAIL setParameters with degradationPreference set should succeed assert_equals: Expect initial param.degradationPreference to be balanced expected (string) "balanced" but got (undefined) undefined
FAIL setParameters with degradationPreference unset should succeed assert_equals: Expect initial param.degradationPreference to be balanced expected (string) "balanced" but got (undefined) undefined
Harness: the test ran to completion.
